+++ ZdorovayaNorka.Common.Entities/Shift.cs
@@ -7,10 +7,10 @@ namespace ZdorovayaNorka.Common.Entities
 {
     public partial class Shift
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateTime StartShiftDate { get; set; }
         public DateTime EndtShiftDate { get; set; }
-        public long? NumberOfHours { get; set; }
-        public long EmployeId { get; set; }
+        public int? NumberOfHours { get; set; }
+        public int EmployeId { get; set; }
     }
 }
